api/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.18
python-jose[cryptography]>=3.4.0
PyJWT>=2.8.0
xmlrpc2>=0.2.0
requests>=2.28.1
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.3.0
python-dotenv>=0.21.0
mistralai>=1.0.0
numpy>=1.24.0
invoice2data>=0.4.0
